+++ src/hooks.js
@@ -1,4 +1,6 @@
-import { useEffect } from 'react';
+import { useState, useEffect } from 'react';
+import StarLayout from '../styles/star-layout'
+
 
 export const useOnClickOutside = (ref, handler) => {
   useEffect(() => {
@@ -16,4 +18,13 @@ export const useOnClickOutside = (ref, handler) => {
   },
   [ref, handler],
   );
-};  
+};
+
+export const useStarTheme = () => {
+  const [theme, setTheme] = useState(starTheme)
+
+  return (
+    {theme === starTheme ? <StarLayout/> : null}
+  )
+
+}
